+++ src/main/java/org/seedstack/coffig/mapper/CompositeMapper.java
@@ -10,14 +10,14 @@ package org.seedstack.coffig.mapper;
 import org.seedstack.coffig.Coffig;
 import org.seedstack.coffig.TreeNode;
 import org.seedstack.coffig.spi.ConfigurationMapper;
-import org.seedstack.coffig.util.AbstractComposite;
+import org.seedstack.coffig.spi.BaseComposite;
 
 import java.lang.reflect.Type;
 import java.util.Arrays;
 
 import static org.seedstack.shed.reflect.Types.rawClassOf;
 
-public class CompositeMapper extends AbstractComposite<ConfigurationMapper> implements ConfigurationMapper {
+public class CompositeMapper extends BaseComposite<ConfigurationMapper> implements ConfigurationMapper {
     private Coffig coffig;
 
     public CompositeMapper(ConfigurationMapper... items) {