+++ src/main/java/org/seedstack/coffig/node/ArrayNode.java
@@ -7,11 +7,10 @@
  */
 package org.seedstack.coffig.node;
 
+import org.seedstack.coffig.TreeNode;
 import org.seedstack.coffig.internal.ConfigurationErrorCode;
 import org.seedstack.coffig.internal.ConfigurationException;
-import org.seedstack.coffig.NamedNode;
 import org.seedstack.coffig.internal.PropertyNotFoundException;
-import org.seedstack.coffig.TreeNode;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -30,7 +29,6 @@ public class ArrayNode extends AbstractTreeNode {
     }
 
     public ArrayNode(ArrayNode other) {
-        super(other);
         this.children = new ArrayList<>(other.children);
     }