+++ src/main/java/org/seedstack/coffig/node/UnmodifiableTreeNode.java
@@ -5,7 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.coffig;
+package org.seedstack.coffig.node;
+
+import org.seedstack.coffig.TreeNode;
 
 import java.util.Optional;
 import java.util.stream.Stream;
@@ -18,11 +20,6 @@ public class UnmodifiableTreeNode implements TreeNode {
             throw new IllegalArgumentException("Null tree node not allowed");
         }
         this.treeNode = treeNode;
-    }
-
-    @Override
-    public NodeAttributes attributes() {
-        return UnmodifiableNodeAttributes.of(treeNode.attributes());
     }
 
     @Override