+++ src/main/java/org/seedstack/coffig/provider/EnvironmentProvider.java
@@ -8,7 +8,7 @@
 package org.seedstack.coffig.provider;
 
 import org.seedstack.coffig.node.MapNode;
-import org.seedstack.coffig.NamedNode;
+import org.seedstack.coffig.node.NamedNode;
 import org.seedstack.coffig.spi.ConfigurationProvider;
 
 public class EnvironmentProvider implements ConfigurationProvider {