+++ src/test/java/org/seedstack/coffig/CoffigTest.java
@@ -13,6 +13,7 @@ import org.seedstack.coffig.fixture.EnumFixture;
 import org.seedstack.coffig.internal.ConfigurationException;
 import org.seedstack.coffig.node.ArrayNode;
 import org.seedstack.coffig.node.MapNode;
+import org.seedstack.coffig.node.NamedNode;
 import org.seedstack.coffig.provider.VoidProvider;
 import org.seedstack.coffig.spi.ConfigurationProvider;