+++ src/test/java/org/seedstack/coffig/mapper/PropertiesMapperTest.java
@@ -10,7 +10,7 @@ package org.seedstack.coffig.mapper;
 import org.junit.Test;
 import org.seedstack.coffig.Coffig;
 import org.seedstack.coffig.node.MapNode;
-import org.seedstack.coffig.NamedNode;
+import org.seedstack.coffig.node.NamedNode;
 import org.seedstack.coffig.spi.ConfigurationMapper;
 
 import java.util.Properties;