+++ src/test/java/org/seedstack/coffig/node/MapNodeTest.java
@@ -10,7 +10,6 @@ package org.seedstack.coffig.node;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.seedstack.coffig.internal.ConfigurationException;
-import org.seedstack.coffig.NamedNode;
 import org.seedstack.coffig.internal.PropertyNotFoundException;
 import org.seedstack.coffig.TreeNode;