+++ src/test/java/org/seedstack/coffig/node/MutableNodesTest.java
@@ -8,7 +8,6 @@
 package org.seedstack.coffig.node;
 
 import org.junit.Test;
-import org.seedstack.coffig.NamedNode;
 import org.seedstack.coffig.internal.PropertyNotFoundException;
 import org.seedstack.coffig.TreeNode;